chore: organize screen actions and their commands
