+++ src/Game/Fleet/Fleet.tsx
@@ -13,8 +13,7 @@ import { shipsAtom } from "../store"
 import { Ship } from "../entities"
 import FleetGrid from "../../components/FleetGrid"
 import { useState } from "react"
-import { useMoveShip } from "../Surface/commands"
-import { useDecommission } from "../../commands"
+import { useMoveShip, useDecommission } from "../actions"
 import ShipIcon from "../../components/ShipIcon"
 import { useSelectedColonizedPlanet } from "../dataHooks"
 import { throwError } from "game-signaling-server/client"
